requirements.txt
# Django and REST Framework
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1

# API Documentation
drf-yasg==1.21.7
setuptools==69.0.0

# Database
psycopg2-binary==2.9.9

# Environment and Configuration
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-django==4.7.0
requests==2.31.0
jsonschema==4.20.0

# Utilities
sqlalchemy==2.0.23
